Move deprecated helper function out of twenty-first crate
It's only used by the `stark_rp` logic, so we can move it to that crate.
